+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added `table.result.Result.ScanNamed()` scan function
 * Changed signature of `table.{Result,StreamResult}` - replace type of result of `NextResultSet` from `bool` to `error`
 * Added `table.{Result,StreamResult}.NextResultSetBool()` method for use with `bool` result
+* Changed connection secure to true by default
 
 ## 3.7.2
 * Retry remove directory in `sugar.RemoveRecursive()` for retryable error